repo-doc-agent/backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangGraph & LangChain
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.15
langchain-groq==0.2.1

# LLM Providers
groq==0.11.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
supabase==2.3.4

# Background Jobs
celery==5.3.6
redis==5.0.1

# Git Operations
GitPython==3.1.41
PyGithub==2.1.1
python-gitlab==4.4.0

# Code Analysis
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.0
radon==6.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1

# Security
cryptography==42.0.0
python-jose[cryptography]==3.3.0

# Monitoring & Logging
loguru==0.7.2
